Extracts handleGetDirections method
