+++ src/components/ResultsPage.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { View, Text, StyleSheet, Linking, Image } from 'react-native';
-import getDirections from 'react-native-google-maps-directions';
+// import getDirections from 'react-native-google-maps-directions';
 import axios from 'axios';
 import renderIf from 'render-if';
 import { Card, CardSection, Button } from './common';
-import { midpoint, placesRating, setTopVenues } from './methods';
+import { midpoint, placesRating, setTopVenues, handleGetDirections } from './methods';
 
 
 class ResultsPage extends Component {
@@ -36,7 +36,7 @@ class ResultsPage extends Component {
     .then(response => { this.setState(midpoint(this.state.p1Latitude, this.state.p1Longitude, this.state.p2Latitude, this.state.p2Longitude)) })
 
     // Google Places API to find places within 500m radius of midPoint => array
-    .then(response => axios.get('https://maps.googleapis.com/maps/api/place/nearbysearch/json?location=' + this.state.lat2 + ',' + this.state.lng2 + '&radius=500&key=AIzaSyByFVMWrXcFmDawtZV1tqvn0fAXgVZe-DY' + '&types=' + this.props.placeType)
+    .then(response => axios.get(`https://maps.googleapis.com/maps/api/place/nearbysearch/json?location=${this.state.lat2},${this.state.lng2}&radius=500&key=AIzaSyByFVMWrXcFmDawtZV1tqvn0fAXgVZe-DY&types=${this.props.placeType}`)
       .then(response => {
         this.setState({ midPlaces: response.data,
                         midPlaceOneId: response.data.results[0].place_id
@@ -44,10 +44,9 @@ class ResultsPage extends Component {
       })
     )
     // Pull ratings from places
-    .then(response => { this.setState(placesRating(this.state.midPlaces.results)) })
-    .then(response => { this.top3RatedArray() })
-    .then(response => {
-      this.setState(setTopVenues(this.state.top3venues));
+    .then(response => { this.setState(placesRating(this.state.midPlaces.results)); })
+    .then(response => { this.top3RatedArray(); })
+    .then(response => { this.setState(setTopVenues(this.state.top3venues));
     });
   }
 
@@ -56,33 +55,28 @@ class ResultsPage extends Component {
     this.setState({ top3venues: topvenues });
   }
 
-  handleGetDirections(lat1, lng1, lat2, lng2) {
-    const data = {
-       source: {
-        latitude: lat1,
-        longitude: lng1
-      },
-      destination: {
-        latitude: lat2,
-        longitude: lng2
-      },
-      params: [
-        {
-          key: 'dirflg',
-          value: 'r'
-        }
-      ]
-    };
-
-    getDirections(data);
-  }
+  // handleGetDirections(lat1, lng1, lat2, lng2) {
+  //   const data = {
+  //      source: {
+  //       latitude: lat1,
+  //       longitude: lng1
+  //     },
+  //     destination: {
+  //       latitude: lat2,
+  //       longitude: lng2
+  //     },
+  //     params: [
+  //       {
+  //         key: 'dirflg',
+  //         value: 'r'
+  //       }
+  //     ]
+  //   };
+  //
+  //   getDirections(data);
+  // }
 
   render() {
-    console.log('lat2 object type:', Object.prototype.toString.call(this.state.lat2))
-    console.log('placesRating (what comes out):', this.state.ratingsArray)
-    console.log('placesRating (what goes in - midPlaces):', this.state.midPlaces)
-
-
     return (
       <Image source={require('../assets/blurryLights.jpg')} style={styles.backgroundImage}>
       <View style={styles.container}>
@@ -105,10 +99,10 @@ class ResultsPage extends Component {
         <CardSection style={styles.cardSection1}>
           <Button
           onPress={() =>
-            this.handleGetDirections(this.state.p1Latitude,
-                                    this.state.p1Longitude,
-                                    this.state.place1lat,
-                                    this.state.place1lng)}
+            handleGetDirections(this.state.p1Latitude,
+                                this.state.p1Longitude,
+                                this.state.place1lat,
+                                this.state.place1lng)}
           >
             Get Directions
           </Button>
@@ -142,10 +136,10 @@ class ResultsPage extends Component {
           <CardSection style={styles.cardSection2}>
             <Button
             onPress={() =>
-              this.handleGetDirections(this.state.p1Latitude,
-                                      this.state.p1Longitude,
-                                      this.state.place2lat,
-                                      this.state.place2lng)}
+              handleGetDirections(this.state.p1Latitude,
+                                  this.state.p1Longitude,
+                                  this.state.place2lat,
+                                  this.state.place2lng)}
             >
               Get Directions
             </Button>
@@ -179,10 +173,10 @@ class ResultsPage extends Component {
           <CardSection style={styles.cardSection3}>
             <Button
             onPress={() =>
-              this.handleGetDirections(this.state.p1Latitude,
-                                      this.state.p1Longitude,
-                                      this.state.place3lat,
-                                      this.state.place3lng)}
+              handleGetDirections(this.state.p1Latitude,
+                                  this.state.p1Longitude,
+                                  this.state.place3lat,
+                                  this.state.place3lng)}
             >
               Get Directions
             </Button>
