+++ src/components/methods/index.js
@@ -1,3 +1,4 @@
 export * from './midpoint';
 export * from './placesRating';
 export * from './setTopVenues';
+export * from './handleGetDirections';
